drm/msm: protect setting fod doze status by allow_bl_update
In the current implementation, doze status command might set the backlight
of the panel before receiving a frame from the driver during the first commit.
There is chance for garbage content to be shown onto the panel.
This change imposes a condition to delay doze status command until the
first frame is received from the HW.