msm/drm: protect nolp entry by allow_bl_update
In the current implementation, nolp entry command might set the backlight
of the panel before receiving a frame from the driver during the first commit.
There is chance for garbage content to be shown onto the panel.
This change imposes a condition to delay nolp entry command until the
first frame is received from the HW.